Move hprof to a separate file
